requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
execnet==2.0.2
h11==0.14.0
idna==3.6
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==23.2
pluggy==1.3.0
pycparser==2.21
PySocks==1.7.1
pytest==7.4.3
pytest-xdist==3.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
testit-adapter-pytest==2.5.3
testit-api-client==3.4.1
testit-python-commons==2.5.3
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.1.0
webdriver-manager==4.0.1
wsproto==1.2.0

pydantic~=2.6.3